Fix for 223 and BCF small fixups: RRR EEE docs on BCF, formatting SSS
Closes # 223
Now user has to select all files needed for a converter, anyway as
seen by the progress bar bug we had no one using BCFs (as we had no
report on that in 2 years), so no need to further complicate the super
complicated BCF code - that is rewrite BCFs (without breaking backwards
compatibility) to somehow not fail if installers with identical content
are present, probably it is not possible at all.
As an added bonus InstallersData.crc_installer finally is binned.

Mopy/bash/bosh/bain.py:
- applyEmbeddedBCFs needs to return a tuple:
- no need to call bass.rmTempDir before installer.unpackToTemp,
unpackToTemp does that

bolt: I run into an edge case - I had the converted auto BCF package
unpacked to a project with identical name and was getting an Access
Denied... comment added but all this sucks

Some typos and wrapping.